requirements.txt
pyspark==3.5.3
pymongo
python-dotenv
flask
flask-cors
numpy
scikit-learn
seaborn
matplotlib